contracts-bedrock: fuzz L2ToL1MessagePasser
Fuzzes the L2ToL1MessagePasser for state changes to its storage.
This specifically tests that the storage slot that the proofs
are generated against is computed correctly. This is important to
test as any changes that break the proof would also break this test.

The fuzz runs were used to generate test cases in
https://github.com/ethereum-optimism/optimism/pull/3469
for off chain computation of the storage slots for withdrawals.